refactor: rcmgr_metrics.go
